Text caching (not really)
In this commit I removed creating new instance of Text class everytime
when text was drawn. Now, when drawing begins, new instance of Text
class is created, and then drawing text is just changing properties of
it and at end of drawing this text is disposed. Simple enought, and I
think it speeded up things a bit. Also, removed empty try of getting
font kerning, what was not used anywhere in MeasureText sub.